+++ client/StrByteBuffer.h
@@ -23,8 +23,8 @@ public:
 	virtual ~StrByteBuffer();
 	virtual unsigned int getBytesLeft();
 protected:
-	virtual unsigned short int readDataInternal(void *destinationBuffer,
-			unsigned short int lenToRead);
+	virtual unsigned int readDataInternal(void *destinationBuffer,
+			unsigned int lenToRead);
 private:
 	std::string str;
 	unsigned int offset;
